backend/requirements.txt
Flask==3.0.3
numpy==1.26.4
openai==0.28.0
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.23
